+++ R/predictions.R
@@ -266,8 +266,8 @@ get_predictions_MixMod <- function(model, fitfram, ci.lvl, linv, type, terms, ty
     if (is.null(prdat.sim) || inherits(prdat.sim, c("error", "simpleError"))) {
       cat(crayon::red("Error: Confidence intervals could not be computed.\n"))
       if (inherits(prdat.sim, c("error", "simpleError"))) {
-        cat(sprintf("* Reason: %s\n", deparse(prdat.sim[[1]])))
-        cat(sprintf("* Source: %s\n", deparse(prdat.sim[[2]])))
+        cat(sprintf("* Reason: %s\n", deparse(prdat.sim[[1]], width.cutoff = 500)))
+        cat(sprintf("* Source: %s\n", deparse(prdat.sim[[2]], width.cutoff = 500)))
       }
 
       fitfram$conf.low <- NA
@@ -940,8 +940,8 @@ get_predictions_glmmTMB <- function(model, fitfram, ci.lvl, linv, type, terms, t
 
         cat(crayon::red("Error: Confidence intervals could not be computed.\n"))
         if (inherits(prdat.sim, c("error", "simpleError"))) {
-          cat(sprintf("* Reason: %s\n", deparse(prdat.sim[[1]])))
-          cat(sprintf("* Source: %s\n", deparse(prdat.sim[[2]])))
+          cat(sprintf("* Reason: %s\n", deparse(prdat.sim[[1]], width.cutoff = 500)))
+          cat(sprintf("* Source: %s\n", deparse(prdat.sim[[2]], width.cutoff = 500)))
         }
 
         fitfram$predicted <- prdat